requirements.txt
undetected-chromedriver==3.5.5
selenium==4.36.0
python-telegram-bot==21.9
aiosqlite==0.20.0
